+++ CHANGELOG.md
@@ -129,3 +129,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### v2.10.0
 - Feat: Installments rules have priorities
 - Feat: Config to no use default installments and use installments rules only, instead
+- Feat: Use customer group to filter installments rules
+- Feat: Use Product Set ID to filter installments rules
+- Feat: Use day of the week for installments rules
+- Feat: Use credit card brand for installments rules
+
+### v2.10.1
+- Fix: error when filtering installments rules by product set